refactor(button): put interface in a separate file
DSY-1866
